Replace Material UI with custom CSS (PostCSS) Removed imports related to `@material-ui` component library to prepare for the migration to custom CSS styles.
Removed theme object and other artifacts related to Material UI.

Restructured JSX to allow for a smoother transition to CSS.

Added a `Board` component that will serve as the central presentation
component for the application. The `Board` component is responsible for
displaying the Kanban board itself, the board's `List` components and
the `Ticket` components that will live inside of the `List`s.